working on update huidigKrediet

+++ bestellingservice/src/main/java/nl/kantilever/bestellingservice/entities/Gebruiker.java
@@ -32,9 +32,9 @@ public class Gebruiker {
 
   private String telefoonnummer;
 
-  private int maxKrediet = 500; // default value
+  private double maxKrediet = 500.0; // default value
 
-  private int huidigKrediet = 0; // default value
+  private double huidigKrediet = 0.0; // default value
 
 
   public Gebruiker(long id, String voornaam, String achternaam) {
